Factored out get_mime_type into util.
